requirements.txt
easydict==1.10
numpy==1.20.3
opencv_python==4.7.0.72
scikit_image==0.18.3
tensorflow_gpu==2.10.0
tf_slim==1.1.0
torch==1.13.1
